LayoutTests/fast/dom/shadow/shadow-root-js-api-expected.txt

Tests for ShadowRoot JS APIs. Can only run within DRT

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowHost is shadowRoot.host
PASS successfullyParsed is true

TEST COMPLETE
